requirements.txt
Pillow>=9.0.0
Nuitka>=1.0.0